+++ R/gorica_methods.R
@@ -22,6 +22,8 @@
 #' @param comparison A character string indicating what the \code{hypothesis}
 #' should be compared to. Defaults to \code{comparison = "unconstrained"};
 #' options include \code{c("unconstrained", "complement", "none")}.
+#' @param iterations Integer. Number of samples to draw from the parameter space
+#' when computing the \code{gorica} penalty.
 #' @param ... Additional arguments passed to the internal function
 #' \code{compare_hypotheses}.
 #'
@@ -112,6 +114,11 @@
 #' \code{estimates}.
 #' \item \code{comparison}  Which alternative hypothesis was used.
 #' \item \code{hypotheses}  The hypotheses evaluated in \code{fit}.
+#' \item \code{relative_weights}  The relative weights of each hypothesis (rows)
+#' versus each other hypothesis in the set (cols). The diagonal is equal to one,
+#' as each hypothesis is equally likely as itself. A value of, e.g., 6, means
+#' that the hypothesis in the row is 6 times more likely than the hypothesis in
+#' the column.
 #' }
 #' @author Caspar van Lissa, Yasin Altinisik, Rebecca Kuiper
 #' @references Altinisik, Y. (2018). Evaluation of Inequality Constrained
@@ -124,18 +131,23 @@
 #'
 #' Kuiper, R. M., Hoijtink, H., & Silvapulle, M. J. (2011).
 #' An Akaike-type information criterion for model selection under inequality
-#' constraints. Biometrika, 98, 495-501. doi:10.1093/biomet/asr002
+#' constraints. Biometrika, 98, 495-501.
+#' <https://academic.oup.com/biomet/article/98/2/495/264032>
 #'
 #' Kuiper, R. M., Hoijtink, H., & Silvapulle, M. J. (2012).
 #' Generalization of the order-restricted information criterion for multivariate
 #' normal linear models. Journal of statistical planning and inference, 142(8),
-#' 2454-2463. \href{https://doi.org/10.1016/j.jspi.2012.03.007}{
-#' doi:10.1016/j.jspi.2012.03.007}
+#' 2454-2463. \doi{10.1016/j.jspi.2012.03.007}
+#'
+#' Vanbrabant, L., Van Loey, N., and Kuiper, R.M. (2019).
+#' Evaluating a theory-based hypothesis against its complement using an AIC-type
+#' information criterion with an application to facial burn injury.
+#' Psychological Methods. \doi{10.1037/met0000238}
 #'
 #' McCullagh, P. & Nelder, J. (1989). Generalized linear models (2nd ed.). Boca
 #' Raton, FL: Chapman & Hall / CRC.
 #'
-#' McCullogh, C. E., & Searle, S. R. (2001). Generalized linear and mixed
+#' McCulloch, C. E., & Searle, S. R. (2001). Generalized linear and mixed
 #' models. New York, NY: Wiley.
 #' @examples
 #' \dontshow{
@@ -148,7 +160,7 @@
 #' gorica(aov1,hypothesis="block1=block5;
 #'    K1<0", iterations = 5)
 #'
-#' # EXAMPLE 3. gml
+#' # EXAMPLE 3. glm
 #' counts <- c(18,17,15,20,10,20,25,13,12)
 #' outcome <- gl(3,1,9)
 #' treatment <- gl(3,3)
@@ -173,7 +185,7 @@
 #' gorica(aov1,hypothesis="block1=block5;
 #'    K1<0")
 #'
-#' # EXAMPLE 3. gml
+#' # EXAMPLE 3. glm
 #' counts <- c(18,17,15,20,10,20,25,13,12)
 #' outcome <- gl(3,1,9)
 #' treatment <- gl(3,3)
@@ -193,7 +205,7 @@
 #' @importFrom stats as.formula coef complete.cases cov lm model.frame
 #' model.matrix pt qt sd setNames summary.lm var vcov
 #'
-gorica <- function(x, hypothesis, comparison = "unconstrained", ...) {
+gorica <- function(x, hypothesis, comparison = "unconstrained", iterations = 100000, ...) {
   UseMethod("gorica", x)
 }
 
@@ -202,6 +214,7 @@ gorica <- function(x, hypothesis, comparison = "unconstrained", ...) {
 gorica.default <- function(x,
                            hypothesis,
                            comparison = "unconstrained",
+                           iterations = 100000,
                            Sigma,
                            ...
 )
@@ -219,44 +232,12 @@ gorica.default <- function(x,
   if(is.list(Sigma) & length(Sigma) == 1) Sigma <- Sigma[[1]]
   names(x) <- rename_function(names(x))
   colnames(Sigma) <- rownames(Sigma) <- names(x)
-  # Parse hypotheses --------------------------------------------------------
-  #ren_estimate <- rename_estimate(x)
-  if(!inherits(comparison, "character")|length(comparison) > 1){
-    stop("Argument 'comparison' must be an atomic character string.")
-  } else {
-    comp_arg <- pmatch(comparison, c("unconstrained", "complement", "none"))
-    if(is.na(comp_arg)) stop("Argument 'comparison' did not match one of the available options: 'unconstrained', 'complement', or 'none'.")
-    comparison <- c("unconstrained", "complement", "none")[pmatch(comparison, c("unconstrained", "complement", "none"))]
-  }
-  if(inherits(hypothesis, "character")){
-    hypothesis <- rename_function(hypothesis)
-    hyp_params <- params_in_hyp(hypothesis)
-    coef_in_hyp <- sort(unique(charmatch(rename_function(hyp_params),
-                             rename_function(names(x)))))
-    if(anyNA(coef_in_hyp)){
-      stop("Some of the parameters referred to in the 'hypothesis' do not correspond to parameter names of object 'x'.\n  The following parameter names in the 'hypothesis' did not match any parameters in 'x': ",
-           paste(hyp_params[is.na(coef_in_hyp)], collapse = ", "),
-           "\n  The parameters in object 'x' are named: ",
-           paste(names(x), collapse = ", "))
-    }
-    if(any(coef_in_hyp == 0)){
-      stop("Some of the parameters referred to in the 'hypothesis' matched multiple parameter names of object 'x'.\n  The following parameter names in the 'hypothesis' matched multiple parameters in 'x': ",
-           paste(hyp_params[coef_in_hyp == 0], collapse = ", "),
-           "\n  The parameters in object 'x' are named: ",
-           paste(names(x), collapse = ", "))
-    }
-    # Drop parameters not in hypothesis
-    x <- x[coef_in_hyp]
-    Sigma <- Sigma[coef_in_hyp, coef_in_hyp]
+  # Perform housekeeping --------------------------------------------------------
 
-    hypothesis <- parse_hypothesis(names(x), hypothesis)
-  } else {
-    if(inherits(hypothesis, "list") & !is.null(hypothesis[["hyp_mat"]]) & !is.null(hypothesis[["n_ec"]])){
-      hypothesis$original_hypothesis <- matrix_to_hyp(hypothesis, names(x))
-    } else {
-      stop("Argument 'hypothesis' must either be a character string with inequality constraints, or a list with an element 'hyp_mat', consisting of a list of hypothesis matrices, and and element 'n_ec', consisting of an integer vector with the number of equality constraints for each hypothesis matrix in 'hyp_mat'.")
-    }
-  }
+  with_env(gorica_housekeeping)
+
+# Evaluate each hypothesis ------------------------------------------------
+
   hypotheses <- mapply(function(this_hyp, nec_num){
     ormle(x,
           Sigma,
@@ -268,6 +249,8 @@ gorica.default <- function(x,
 
   hyp <- reverse_rename_function(hypothesis$original_hypothesis)
 
+# Add unconstrained hypothesis --------------------------------------------
+
   if(comparison == "unconstrained"){
     hypotheses <- c(hypotheses,
                     list(ormle(est = x,
@@ -278,18 +261,45 @@ gorica.default <- function(x,
                     ))
     hyp <- c(hyp, "Hu")
   }
-  res <- compare_hypotheses(hypotheses, ...)
+
+
+# Add complement ----------------------------------------------------------
+
+  Args_res <- c(
+    list(
+      object = hypotheses,
+      iterations = force(iterations)
+      ),
+    list(...)
+  )
+  res <- do.call(compare_hypotheses, Args_res)
   fit <- res$comparisons
 
   if(comparison == "complement"){
-    complement <- do.call(comp, c(hypotheses[[1]], wt_bar = res[[1]][[2]]))
-    fit <- rbind(fit, complement)
-    hyp <- c(hyp, "Hc")
+    if(length(res[["gorica_penalties"]]) > 1){
+      warning("Cannot compute complement for more than one hypothesis.", call. = FALSE)
+    } else {
+      use_wtbar <- res[["gorica_penalties"]][[1]][["wt_bar"]]
+      use_wtbar <- use_wtbar[length(use_wtbar) - hypothesis$n_ec[1]]
+      complement <- do.call(comp, c(hypotheses[[1]], wt_bar = use_wtbar))
+      fit <- rbind(fit, complement)
+      hyp <- c(hyp, "Hc")
+    }
   }
 
+
+# Prepare output ----------------------------------------------------------
+
+  if(any(fit$penalty < 0)) warning("Some gorica penalties were below 0. This is not theoretically possible. Please send a bug report to c.j.vanlissa@uu.nl", call. = FALSE)
   fit$gorica_weights <- compute_weights(fit$gorica)
 
   Goricares[c("fit", "hypotheses")] <- list(fit, hyp)
+
+  Goricares$relative_weights <- Goricares$fit$gorica_weights %*% t(1/Goricares$fit$gorica_weights)
+  mat_nams <- hyp
+  mat_nams[!nchar(mat_nams) == 2] <- paste0("H", 1:sum(!nchar(mat_nams) == 2))
+  #browser()
+  colnames(Goricares$relative_weights) <- rownames(Goricares$relative_weights) <- mat_nams
   class(Goricares) <- "gorica"
   Goricares
 }
@@ -302,6 +312,7 @@ gorica.htest <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
+           iterations = 100000,
            ...) {
     stop("To be able to run gorica on the results of an object returned by t.test(), you must first load the 'gorica' package, and then conduct your t.test. The standard t.test does not return group-specific variances and sample sizes, which are required by gorica. When you load the gorica package, the standard t.test is replaced by a version that does return this necessary information.")
   }
@@ -312,11 +323,13 @@ gorica.t_test <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
+           iterations = 100000,
            ...) {
     cl <- match.call()
     Args <- as.list(cl[-1])
-
-    Args$x <- x$estimate
+    ests <- get_estimates(x)
+    Args$x <- ests$estimate
+    Args$hypothesis <- force(hypothesis)
     #Args$n <- x$n
 
     if(length(x$estimate) == 1){
@@ -351,6 +364,7 @@ gorica.lm <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
+           iterations = 100000,
            ...) {
 
     cl <- match.call()
@@ -362,7 +376,7 @@ gorica.lm <-
     }
     Args$x <- coef(x)
     Args$Sigma <- vcov(x)
-
+    Args$hypothesis <- force(hypothesis)
     Gorica_res <- do.call(gorica, Args)
     Gorica_res$call <- cl
     Gorica_res$model <- x
@@ -380,6 +394,7 @@ gorica.mplus.model <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
+           iterations = 100000,
            ...) {
 
     cl <- match.call()
@@ -387,7 +402,7 @@ gorica.mplus.model <-
     mplus_est <- get_estimates(x)
     Args$x <- mplus_est$estimate
     Args$Sigma <- mplus_est$Sigma
-
+    Args$hypothesis <- force(hypothesis)
     Gorica_res <- do.call(gorica, Args)
     Gorica_res$call <- cl
     Gorica_res$model <- x
@@ -401,12 +416,16 @@ gorica.mplus.model <-
 
 #' @method gorica lavaan
 #' @export
+#' @rdname gorica
+#' @param standardize Logical. For \code{lavaan} objects, whether or not to
+#' extract the standardized model coefficients. Defaults to \code{FALSE}.
 gorica.lavaan <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
-           standardize = FALSE,
-           ...) {
+           iterations = 100000,
+           ...,
+           standardize = FALSE) {
     cl <- match.call()
     Args <- as.list(cl[-1])
     mplus_est <- get_estimates(x, standardize)
@@ -430,6 +449,7 @@ gorica.lmerMod <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
+           iterations = 100000,
            ...) {
 
     cl <- match.call()
@@ -437,7 +457,7 @@ gorica.lmerMod <-
 
     Args$x <- fixef(x)
     Args$Sigma <- vcov(x)
-
+    Args$hypothesis <- force(hypothesis)
     Gorica_res <- do.call(gorica, Args)
     Gorica_res$call <- cl
     Gorica_res$model <- x
@@ -455,6 +475,7 @@ gorica.model_estimates <-
   function(x,
            hypothesis,
            comparison = "unconstrained",
+           iterations = 100000,
            ...) {
 
     cl <- match.call()
@@ -462,7 +483,7 @@ gorica.model_estimates <-
 
     Args$x <- x$estimate
     Args$Sigma <- x$Sigma
-
+    Args$hypothesis <- force(hypothesis)
     Gorica_res <- do.call(gorica, Args)
     Gorica_res$call <- cl
     Gorica_res$model <- x
@@ -473,4 +494,215 @@ gorica.model_estimates <-
     class(Gorica_res) <- c("gorica_model_estimates", class(Gorica_res))
     Gorica_res
   }
+
+
+#' @section Contingency tables:
+#' When specifying hypotheses about contingency tables, the asymptotic
+#' covariance matrix of the model estimates is derived by means of
+#' bootstrapping. This makes it possible for users to define derived parameters:
+#' For example, a ratio between cell probabilities. For this purpose, the
+#' \code{\link[bain]{bain}} syntax has been enhanced with the command \code{:=}.
+#' Thus, the syntax \code{"a := x[1,1]/(x[1,1]+x[1,2])"} defines a new parameter
+#' \code{a} by reference to specific cells of the table \code{x}. This new
+#' parameter can now be named in hypotheses.
+#' @rdname gorica
+#' @method gorica table
+#' @export
+#' @importFrom limSolve ldei
+gorica.table <- function(x,
+                         hypothesis,
+                         comparison = "unconstrained",
+                         ...){
+  #browser()
+  cl <- match.call()
+  Args <- as.list(cl[-1])
+  original_estimate <- x
+
+  # Get constraints from hypothesis
+  n_pars <- length(original_estimate)
+
+  const <- paste0(get_const(hypothesis), collapse = ";")
+  hypothesis <- rename_table_est(paste0(get_hyp(hypothesis), collapse = ";"))
+  coefs_in_hyp <- params_in_hyp(hypothesis)
+  est <- do.call(get_estimates, list(x = force(x),
+                                        constraints = const))
+  x <- est$estimate
+  Sigma <- est$Sigma
+
+  names(x) <- rename_table_est(rename_function(names(x)))
+  colnames(Sigma) <- rownames(Sigma) <- names(x)
+
+  # Put x and Sigma in Args already, just to have an unaltered backup
+  Args$x <- x
+  Args$Sigma <- Sigma
+
+  # Perform housekeeping --------------------------------------------------------
+
+  with_env(gorica_housekeeping)
+  #with_env(drop_empty_hyps)
+  null_hyps <- sapply(hypothesis$hyp_mat, is.null)
+  if(any(null_hyps)){
+    which_null <- which(null_hyps)
+    remove_these <- as.vector(sapply(which_null, seq_tuples, tuples = 2))
+    hypothesis$hyp_mat[which_null] <- NULL
+    hypothesis$n_constraints <- hypothesis$n_constraints[-remove_these]
+    hypothesis$n_ec <- hypothesis$n_ec[-which_null]
+    hypothesis$original_hypothesis <- hypothesis$original_hypothesis[-which_null]
+  }
+
+# Specific solutions for gorica.table -------------------------------------
+  if(!const == "" & any(x == 1 | x == Inf)) {
+    stop("Some of the defined parameters are invalid (with values equal to 1 or infinity) due to empty cell(s) in the table. Please rewrite the hypotheses.")
+  }
+  if(all(Sigma == 0)){
+    stop("All parameter covariances are equal to zero.")
+  }
+
+# 0. Check if eta's sum to 1
+raw_probs <- sum(x) == 1
+# 1. Check singular covariance matrix. If so, linear dependency and:
+linear_dependency <- !pos_definite(Sigma)
+if(linear_dependency){
+  if(!const == ""){
+    stop("The defined parameters are linearly dependent on each other. Consequently, their covariance matrix is not positive definite. Please rewrite the hypotheses.")
+    }
+  # 2. Check eta’s without variation.
+  null_coefs <- rowSums(Sigma == 0) == ncol(Sigma)
+  # If any:
+  # 3. Check if this/these eta == 0.
+  if(any(null_coefs)){
+    if(any(x[null_coefs] != 0)){ # So, at least one of the eta’s without variation is non-zero
+
+      # If not, then message about rewriting.
+      stop("Please rewrite the hypothesis.")
+    } else {
+      # If so:
+      # Delete eta(s) without variation (which are now all 0; which gives eta_adj);
+      # delete corresponding rows and columns in Sigma (which gives Sigma_adj)
+      # and corresponding columns in R (R_adj);
+      zero_est <- which(null_coefs)
+
+      # Discard the rows and thus columns from Sigma for which all elements are zero.
+      # discard the corresponding etas which leads to etaadj
+      # and the corresponding columns from the restriction matrix Rm, which leads to Rmadj
+      # and do the check to see if you need to adjust the rhs (which then gives rhs_adj)
+      #with_env(hypothesis_remove_nulls, which_par = zero_est)
+      R <- do.call(rbind, hypothesis$hyp_mat)
+      if(raw_probs){
+        remove_par <- max(which(x != 0))
+        R <- sweep(R, MARGIN = 1, as.vector(R[, remove_par]))
+      }
+      rhs <- R[, ncol(R), drop = FALSE]
+      R <- R[, -ncol(R), drop = FALSE]
+
+      # Delete eta(s) without variation (which are now all 0; which gives eta_adj);
+      x <- x[-zero_est]
+      # delete corresponding rows and columns in Sigma (which gives Sigma_adj)
+      Sigma <- Sigma[-zero_est, -zero_est]
+
+      # and the corresponding columns from the restriction matrix Rm, which leads to Rmadj
+      R_adj <- R[, -zero_est, drop = FALSE]
+
+      # and do the check to see if you need to adjust the rhs (which then gives rhs_adj)
+      null_rows <- which(apply(R_adj, 1, function(i){all(i == 0)}))
+      if(length(null_rows) > 0){
+        if(any(rhs[null_rows] > 0)){
+          qadj <- which(apply(R[null_rows, , drop = FALSE], 2, function(j){any(j != 0)}))
+
+          # Fix below
+          G <- diag(1,ncol(hypothesis$hyp_mat[[1]])-1)[null_rows,]
+          H <- rep(0,length(null_rows))
+          q <- ldei(E=R, F=rhs, G = G, H = H)$X
+          q[qadj] <- 0
+          # Overwrite rhs
+          rhs <- R%*%q
+
+        }
+      }
+      new_hypmat <- cbind(R_adj, rhs)
+      lengths <- c(0, sapply(hypothesis$hyp_mat, nrow))
+      hypothesis$hyp_mat <- lapply(1:length(hypothesis$hyp_mat), function(mat_num){
+        new_hypmat[(sum(lengths[1:mat_num])+1):sum(lengths[1:(mat_num+1)]), ]
+      })
+      #with_env(hypothesis_sums_to_one)
+
+    }
+  }
+}
+
+# Then, check whether there is still linear dependency; i.e., check det(Sigma_adj) == 0.
+# If not, then proceed with eta_adj, Sigma_adj, R_adj, and rhs or rhs_adj.
+# If so, check whether the sum of the (remaining) eta’s (i.e., sum(eta_adj) is 1
+if(!pos_definite(Sigma)){
+  if(sum(x) == 1){
+    # If so, rewrite the hypothesis by setting last column of remaining eta’s to 1 – rest; etc; but now using eta_adj, Sigma_adj, R_adj, and rhs or rhs_adj.
+    message("The hypotheses involve all table cells, which introduces a linear dependency. The final cell probability was defined as one minus the other cell probabilities, and hypotheses were respecified to reflect this.")
+    remove_par <- max(which(x != 0))
+    # If hyp sum to one, remove last non-zero par
+    hypothesis$hyp_mat <- lapply(hypothesis$hyp_mat, function(R){
+      sweep(R[, -remove_par, drop = FALSE], MARGIN = 1, as.vector(R[, remove_par]))
+    })
+    # Drop parameters not in hypothesis
+    x <- x[-remove_par]
+    Sigma <- Sigma[-remove_par, -remove_par]
+
+  } else {
+    # If not, then message about rewriting.
+    stop("Please rewrite the hypothesis.")
+  }
+}
+
+#
+# I am not sure whether you need to check for 1 (and further again) after 3 or 4.
+
+
+  # if(all(x[coefs_in_hyp] == 0)){
+  #   print(out)
+  #   stop("All estimates referenced in the hypothesis are equal to zero.")
+  # }
+  # Remove null coefficients ------------------------------------------------
+
+  # if(any(null_coefs)){
+  #   browser()
+  #   #remove_par <- max(which(eta == 0))
+  #   zero_est <- which(null_coefs)
+  #   # Discard the rows and thus columns from Sigma for which all elements are zero.
+  #   # discard the corresponding etas which leads to etaadj
+  #   # and the corresponding columns from the restriction matrix Rm, which leads to Rmadj
+  #   with_env(hypothesis_remove_nulls, which_par = zero_est)
+  #   #hyp <- matrix(c(1,0,0,1,2,1), nrow = 2, byrow = T)
+  #   with_env(hypothesis_sums_to_one)
+  # }
+  #
+  # if(!const == ""){
+  #   if(!is.positive.definite(Sigma)){
+  #     stop("The defined parameters are linearly dependent on each other. Consequently, their covariance matrix is not positive definite. Please rewrite the hypotheses.")
+  #   }
+  # } else {
+  #   if((1 - sum(x)) < .0001){ browser()} # Watch out: Maybe bootstrapped estimates do not exactly sum to 1 in the case of linear dependency
+  #   if(sum(x) == 1){
+  #
+  #   }
+  # }
+  #
+
+
+
+
+#if(any())
+  Args$x <- x
+  Args$Sigma <- Sigma
+  Args$hypothesis <- hypothesis
+  Args$comparison <- force(comparison)
+  Gorica_res <- do.call(gorica, Args)
+  Gorica_res$estimates <- est$estimate
+  Gorica_res$Sigma <- est$Sigma
+  Gorica_res$call <- cl
+  Gorica_res$model <- original_estimate
+  Gorica_res$hypotheses <- sapply(Gorica_res$hypotheses, reverse_rename_table_est)
+
+  class(Gorica_res) <- c("gorica_table", class(Gorica_res))
+  return(Gorica_res)
+}
+