+++ NEWS.md
@@ -1,3 +1,11 @@
+# gorica 0.1.1
+
+* Fixed gorica.t_test as per Prof. dr. Brian Ripley's notification
+* Bugfix to complement
+* Add warnings if penalty terms are < 0
+* Add warning if user tries to compute complement for more than 1 hypothesis
+* Add Leonard Vanbrabant as contributor
+
 # gorica 0.1.0
 
 * New submission